+++ pkg/middleware/component_batch_middleware.go
@@ -5,6 +5,7 @@ import (
 	"encoding/json"
 	"errors"
 	"github.com/grpc-ecosystem/go-grpc-middleware/logging/zap/ctxzap"
+	"github.com/scanoss/go-component-helper/componenthelper"
 	"github.com/scanoss/papi/api/commonv2"
 	"scanoss.com/licenses/pkg/dto"
 	"strings"
@@ -15,8 +16,8 @@ type ComponentBatchMiddleware[TOutput any] struct {
 	MiddlewareBase
 }
 
-func NewComponentsRequestMiddleware(req *commonv2.ComponentsRequest, ctx context.Context) Middleware[[]dto.ComponentRequestDTO] {
-	return &ComponentBatchMiddleware[[]dto.ComponentRequestDTO]{
+func NewComponentsRequestMiddleware(req *commonv2.ComponentsRequest, ctx context.Context) Middleware[[]componenthelper.ComponentDTO] {
+	return &ComponentBatchMiddleware[[]componenthelper.ComponentDTO]{
 		MiddlewareBase: MiddlewareBase{s: ctxzap.Extract(ctx).Sugar()},
 		req:            req,
 	}
@@ -44,34 +45,24 @@ func (m *ComponentBatchMiddleware[TOutput]) groupComponentsByPurl(c []dto.Compon
 	return componentMap
 }
 
-func (m *ComponentBatchMiddleware[TOutput]) Process() ([]dto.ComponentRequestDTO, error) {
+func (m *ComponentBatchMiddleware[TOutput]) Process() ([]componenthelper.ComponentDTO, error) {
 
 	if len(m.req.GetComponents()) == 0 {
 		m.s.Warn("No components request data supplied to decorate. Ignoring request.")
-		return []dto.ComponentRequestDTO{}, errors.New("no components request data supplied")
+		return []componenthelper.ComponentDTO{}, errors.New("no components request data supplied")
 	}
 
 	data, err := json.Marshal(m.req.GetComponents())
 	if err != nil {
 		m.s.Errorf("Problem marshalling dependency request input: %v", err)
-		return []dto.ComponentRequestDTO{}, errors.New("problem marshalling request input data")
+		return []componenthelper.ComponentDTO{}, errors.New("problem marshalling request input data")
 	}
 
-	var componentEntity []dto.ComponentRequestDTO
-	err = json.Unmarshal(data, &componentEntity)
+	var componentDTO []componenthelper.ComponentDTO
+	err = json.Unmarshal(data, &componentDTO)
 	if err != nil {
 		m.s.Errorf("Parse failure: %v", err)
-		return nil, errors.New("failed to parse request input data")
+		return []componenthelper.ComponentDTO{}, errors.New("failed to parse request input data")
 	}
-
-	componentMap := m.groupComponentsByPurl(componentEntity)
-
-	// Convert map to slice
-	components := make([]dto.ComponentRequestDTO, 0, len(componentMap))
-	for _, component := range componentMap {
-		components = append(components, component)
-	}
-	m.s.Debugf("components: %v\n", components)
-	m.s.Debugf("%v to process\n", len(components))
-	return components, nil
+	return componentDTO, nil
 }